+++ docs/release-notes.md
@@ -1,5 +1,7 @@
 ## 0.8.0
 * Added download form for `input-format` and `predict-test` endpoints
+* Changed HTTP GET on `input-format` and `predict-test` to only return HTML
+* Added HTTP POST on `input-format` and `predict-test` which will return JSON
 * Removed `starlette_kwargs` and `uvicorn_kwargs` from `run_model_serving` and
 `run_pandas_serving`